# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  httpErrors.RestErr:
    type: object
...
  models.Asset:
    properties:
      amount:
        type: number
      asset_id:
        type: string
      created_at:
        type: string
      market_id:
        type: string
      notation_at:
        type: string
      portfolio_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    required:
    - amount
    - market_id
    - portfolio_id
    - quantity
    type: object
  models.AssetAdd:
    properties:
      amount:
        type: number
      marketId:
        type: string
      notationAt:
        type: string
      portfolio_id:
        type: string
      quantity:
        type: integer
    required:
    - amount
    - marketId
    - portfolio_id
    - quantity
    type: object
  models.AssetBase:
    properties:
      amount:
        type: number
      asset_id:
        type: string
      image_url:
        type: string
      notation_at:
        type: string
      portfolio_id:
        type: string
      quantity:
        type: integer
      ticker:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - amount
    - portfolio_id
    - quantity
    - ticker
    - title
    - user_id
    type: object
  models.AssetList:
    properties:
      assets:
        items:
          $ref: '#/definitions/models.AssetBase'
        type: array
      has_more:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.AssetTotal:
    properties:
      averagePurchasePrice:
        type: number
      firstNotationAt:
        type: string
      identify:
        type: string
      image_url:
        type: string
      marketId:
        type: string
      ticker:
        type: string
      title:
        type: string
      totalAmount:
        type: number
      totalCount:
        type: integer
      totalQuantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API
  title: REST API
  version: "1.0"
paths:
  /asset/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: asset_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Удалить актив
      tags:
      - Asset
    get:
      consumes:
      - application/json
      parameters:
      - description: asset_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Asset'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Получить актив по идентификатору
      tags:
      - Asset
    put:
      consumes:
      - application/json
      parameters:
      - description: asset_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Asset'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Обновить данные актива
      tags:
      - Asset
  /asset/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add asset
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AssetAdd'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Asset'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Создать новый актив пользователя
      tags:
      - Asset
  /asset/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AssetList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Активы пользователя
      tags:
      - Asset
  /asset/market/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: market_id
        in: path
        name: id
        required: true
        type: integer
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssetList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Получить все активы по идентификатору ценной бумаги/товара
      tags:
      - Asset
  /asset/total-all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.AssetTotal'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestErr'
      security:
      - Auth: []
      summary: Активы пользователя, подсчитанные
      tags:
      - Asset
...
securityDefinitions:
  Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"